chore: install + upgrade deps, pull out scripts, add instructions
